+++ main.py
@@ -7,7 +7,7 @@ Author: RimWorld Linux Community
 License: MIT
 """
 
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
 import sys
 import os